renamed dashboard controllers to home
